Add option to prevent empty hyps from being considered terminated.
Epsilon-emitting models like RNN-T have a propensity to emit empty (all-epsilon) hyps even in the presence of speech because epsilons are low cost and safe.  At decoding time this is generally not a problem if the empty hyp is not the top 1.  However, in minimum word error rate training, it becomes problematic because we are using all hyps from the beam to compute loss.

This CL adds the option to simply not consider the all-epsilon hyp as terminated, so it never gets passed on to the TopKTerminatedHypsOp.  Instead it just gets thrown out.

Note that a hyp that terminates in EOS is not considered empty, so this flag has no effect for non-epsilon-emitting models.

PiperOrigin-RevId: 216612423