Remove RecordYielder::current_epoch.
This is removed from the base class as it's hard to maintain clear and consistent meaning for current_epoch across BasicRecordYielder and WeightedMixRecordYielder. Note that we keep BasicRecordYielder::current_epoch which has clear meaning. RecordYielder::current_epoch is only used in logging now and was added specifically for that so we don't lose much by removing it.

PiperOrigin-RevId: 225385960