Adds a variant of Xavier initialization that uses geometric mean, instead of arithmetic mean, of fan_in and fan_out.
Consider the case of low-rank factorization: two matrices A and B, of shape (m, n) and (n, m), whose product C is a (m, m) matrix. We want Var(C) to be 1/m according to equation (10) and (11) in the Xavier paper. This can be satisfied when Var(A) and Var(B) are both 1/sqrt(m*n). From this perspective, a uniform distribution with scale sqrt(3/sqrt(m*n)) makes more sense than one with scale sqrt(6/(m+n)).

PiperOrigin-RevId: 233442855