Add property to cluster to fetch total worker devices.
Currently, num_devices is switched on the job type, but in many cases
you want to know the number of worker devices regardless of what
the current job type is: e.g., computing the total effective batch size.

This allows all jobs to agree on the total number of workers in the
cluster.

PiperOrigin-RevId: 237381737