Make 'attention_plot_font_properties' a string parameter.
In the original change, this was a string parameter, and it is still
interpreted as a string parameter in the code (it's passed to fname).
But the default value of this parameter is not currently a string.
In running this in colab, it appears that the default branch of this
code (where the parameter is not set) does not work.

In this change we make it clear that this is a string filename
parameter, but we make a default FontProperties object if the string
is empty.

PiperOrigin-RevId: 251368787