Add new BaseTask API to allow customization of final decoding phase.
Background: Currently the implementor of a task controls the
Decode() and PostProcessDecode() functions, which operate on single
examples.  For any sort of work that occurs at the very end of
decoding on all examples, the framework only provides one
mechanism for user-controlled code to be called: via the Summary()
API of a Metric.

While a lot can be implemented via this API, one aspect that is missing
is the ability to write data to the filesystem, because no path
is provided to metrics that gives it a location to write.  Although
we could simply give a model access to the decoder path, it seems more
extensible to allow a more generic API that allows a user to override
an implementation of a finalization method.

We thus implement DecodeFinalize() in BaseTask() whose default implementation
is unchanged (writes key value pairs) but whose design gives future flexibility
to task owners to implement a finalization mechanism of their choice.

The function takes a namedtuple (to allow for future extensibility in the
arguments it can take).  Implementors can both customize the work done
in this API, or do further work as a 'final phase' when decoding each
checkpoint.

PiperOrigin-RevId: 261962824