Provides a Params based description of DataSources as an alternative to p.file_pattern.
file_pattern is currently very flexible, allowing a user to specify either a string (to describe a single file), a comma separated string (to specify multiple files), a list of (file_pattern, weight) pairs or a list of (file_pattern, weight, bprop_variable_filter) 3-tuples. While very useful, it not readable.

This creates DataSources as a separate mechanism to provide a file_pattern and its associated options.

This change is backward compatible with the current file_pattern format.

PiperOrigin-RevId: 263231332