Add helpful utilities / factories for beam-based pipelines to lingvo.
Adds beam to dockerfile as well.

Beam-based pipelines can be used for a variety of things.  The predominant
need is for running data generation pipelines in parallel, but we have also
found it useful for running inference / decoding over large-scale datasets.

This provides some helpful utilities that abstract the underlying implementations
and makes them 'pluggable'.

PiperOrigin-RevId: 263245378