Fix the ToText/FromText conversion when a param type is either a string or a list.
This is a quite common pattern. For instance, layers.FeedForwardNet has `activation` as a param which can be a string or a list of strings. If list of strings is used, it will be encoded as a string in text representation. In such a case FromText may not be able to understand the value should be a list as the default param value is a string itself.

PiperOrigin-RevId: 263413552