Creates a new DataSource class, CurriculumDataSource, to support curriculum learning ("Curriculum Learning", Bengio et al. 2009) and other more complicated ways of scheduling the presentation of training examples.
The parameters define child DataSources that are used to generate data during each training stage.  The boundaries of the training stages are defined in terms of global_step.  This is consistent with the semantics of learning rate scheduling boundaries as defined in schedule.PiecewiseSchedule.  This makes it easy (but optional) to synchronize learning rate stages and training data stages.

PiperOrigin-RevId: 264233276