tasks/car/base_extractor: Change preprocessors to a hyperparams.Params().
This ensures that the root params object can be printed nicely, preprocessors
params entries can be overridden by model_params_override.

This requires using p.Define() when adding to an existing params, rather
than just overridiing the key, which is where the bulk of changes are.
I added a temporarily helper routine to map from NestedMap to Params
but those can be individually removed later.

PiperOrigin-RevId: 267271011