Create pip package files for lingvo.
The eventual goal is to be able to do pip[3] install lingvo[-gpu]
so that lingvo can be used as a library.  More work is needed to get
to this point, but having an installable pip package is one step
towards that.

This also enables using lingvo in Cloud Dataflow pipelines.

PiperOrigin-RevId: 268119475