Implements GraphStep, StackStep, and StatelessLayerStep.
GraphStep is like GraphLayer, but extended to handle Steps.
It does the state management and external input management for
each layer transparently.

StackStep is like LayerStack. It's a simple way of combining multiple
steps into a stack structure, and can work well when the connections
between steps aren't complicated.

StatelessLayerStep is a simple wrapper for layers that don't have
recurrent state.

PiperOrigin-RevId: 268541011