Change tf.to_int32(tf.reduce_sum( to tf.to_int32(tf.round(tf.reduce_sum.
There is no reason to believe there was any problems with the previous
implementation. With paddings being only 0s and 1s the sum should be exact and
the round should be unnecessary. However, just in case there were any problems,
this should fix it.

This may also preempt any problems that may occur when using quantization/low precision.

PiperOrigin-RevId: 269892522