Add a test utility to easily test that a registered model builds.
In contrast to models_test_helper, which only verifies that the params
can be instantiated to create a model, this utility goes one step
further and tries to construct the Trainer / Decoder for a model,
which requires doing some more work and API calls of the model.

This is useful if one wants to run this test locally before launching
a job on a real cluster, because it quickly highlights issues that
might show up when running a full trainer and decoder.

Notably, this does not run any steps of training.

PiperOrigin-RevId: 285817356