Enable target_sequence_sampler for Transformer decoder
The previous implementation actually works on TPUs, because in that case the beam search states all have fixed shapes. However when decoding happens on CPU/GPUs (which is usually the case for decoder jobs), the shape dim dynamically increase, which is in conflict with recurrent.py based on which target_sequence_sampler is implemented. This CL basically fix these cases by padding state values to fixed shapes. These changes however are only based on MT decoder implementations.

PiperOrigin-RevId: 286731166