Make InputBatch/PreprocessInputBatch protected functions. The public API has always been GetPreprocessedInputBatch, and that is the input batch actually used in training. Tests that use InputBatch() inputs may not reflect the actual inputs used by the model.
Remove base_model.GetInputBatch. It's very specifically for distributed execution/TPU infeed and is not a generic approach for getting an input batch for unit tests. The very generic method name is probably the reason for many people using it for unit tests. (while it's safe to use it for unit tests if we end up changing the functionality in the future it could break many tests which would be annoying)

PiperOrigin-RevId: 303883293