Introduces a "single-dict" restriction to DefineTFDataInput.
This CL constrains DefineTFDataInput to accept only tf.data functions that yield a dict or its subclasses (including Lingvo's NestedMap), and let generated input generators return a NestedMap converted from the original dict.
This restriction is currently necessary to maintain several consistency with some functionalities on BaseInputGenerator that rely on NestedMap.
And this may be reasonable so far, because most non-trivial tf.data pipelines return a dict, such as datasets in the TensorFlow Datasets repository.

PiperOrigin-RevId: 320104989