Implement a decoder for the image classifier.
For now, just output whether the example was correctly
in the top 1 or top 5.

In the future, more kinds of 'analysis' that isn't
already served by Inference() can be done here.

It is possible to simply output the logits
for every example and let the top-k analysis be done
offline, but given that the outputs here are very small
it seems harmless to compute it in TF.

PiperOrigin-RevId: 327362966