Add a tool to count the number of tfrecords in a file pattern.
TFRecords don't store metadata about the number of records in
the header/footer, requiring that we scan over the entire
dataset.

Although it is possible to write a tool to read each file
in parallel to count the records (or use the C++ API to
accomplish this more efficiently), it's often easier to just
launch a beam job that can do this quickly on a bunch of
machines in parallel.

Hence, this script.

PiperOrigin-RevId: 328220285