Reduce default bucket_upper_bound from 1000000000 to 1000000.
An int64 histogram is created in record_batcher.h to record the
distribution of examples into buckets for the purposes of understanding
bucket distributions.  Its size is created by looking at the last element
of bucket_upper_bound and creating a bin for every possible bucket value.

If a user does not configure bucket_upper_bound in a custom way, the job
will end up allocating an 8GiB memory region.

This CL reduces the value by a factor of 1000 to reduce this memory region;
although the ideal default value probably should have been 1, we don't know
if there are some unintended cases where someone is specifying buckets > 1
and relying on the fact that the bucket upper bound is insanely high.  It
seems less likely that someone is producing bucket values > 1000000 and
expecting this default to never change.  The answer in that case is to have
the end-user customize the bucket upper bound anyway.

PiperOrigin-RevId: 332867672