BaseExtractor: Filter out keys from other extractors.
Prior to this change, extractors' _Extract() functions had unintended
access to the parsed features of other extractors.  The API is
supposed to be that an extractor has access to the features it has
extracted itself.

This change filters the feature dictionary prior to passing
to each Extractor's _Extract() function.

Should one need to have access to another extractor's output, that
functionality should live in a preprocessor, where explicit ordering
and mixing of data is intended.

Adds a test to verify this behavior.

PiperOrigin-RevId: 333187251