Allow merging select CPU inputs with the output of Decode() for TPU decoding.
Some decoding functions might want inputs that cannot or do not need
to be sent to the TPU device on every step.  For example, some small
string tensors (e.g., the key of the example) may be useful for logging
in a decoder summary for debugging purposes.  Alternatively, some outputs
needed for visualization simply need to be passed through to
PostProcessDecodeOut, rather than streaming it untouched through
the infeed/outfeed of a TPU.

This CL adds the analog equivalent of a 'queue' for CPU tensors
that should be stored and dequeued along with the outputs of the
Decode() function on TPU, and then merged together to pass
to PostProcessDecodeOut.

For the car task, update the base_extractor code to specify
string tensors in the input as CPU passthrough values by default.

The TpuEnqueue function now filters out CPU passthrough tensors,
so we no longer need to filter out string tensors in base_extractor.

Backwards compat:
- By default, CPU passthrough keys is empty, so nothing gets
  filtered, nothing gets enqueued, dequeue returns None so nothing
  additionally gets requested.

- In car codebase, CPU passthrough keys are string tensors.
  Non-TPU workloads don't treat these specially, and TPU workloads
  using TPUEnqueue in lingvo will filter them out equivalent
  to the filtering we did in _NestedMapFromBatchedOutputs.
  Non-lingvo use cases typically don't use _NestedMapFromBatchedOutputs
  and so are responsible for doing this filtering themselves already.

PiperOrigin-RevId: 333759122