Support GenericInput op called from a tf.data.Dataset
Creates a self-contained function library definition for the cloned library runtime.
This fixes the issue where, when called from a tf.data.Dataset, the cloned library
runtime wouldn't contain the user-provided function, which would lead to a
NotFoundError. With this change, GenericInput can be used in combination with
tf.data.

PiperOrigin-RevId: 334474539