Track step seeds separately by graph, each initialized to a different random value.
Add an option propagate_global_tensors to py_utils.Function that wraps the functions with step seed when set.

This eliminates the need to pass step seeds in and out of graphs, as there is no
need for step seeds to be consecutive at graph boundaries, only that there is
not a repeated seed used.

PiperOrigin-RevId: 338894989