Program abstraction: Compile all programs in parallel.
Since TPU compilation can take minutes to run, we
compile each program in parallel, so that we overlap
the compilation of the various programs in the schedule.

This reduces the overall latency to the first step of training
by ensuring that the latency of the compile step is
just the maximum compile time of any program, rather than
the sum.

PiperOrigin-RevId: 339746364