Add interface for batched preprocessing.
In some cases, we need to process batches of data using the extractor
framework, rather than single examples.  This provides an interface
for extractors and preprocessors that allows more efficient, batched
versions of the implementations to be specialized.

The default implementation does a simple loop over each example,
calling the existing single-example implementation, but one can
specialize the implementation (as demonstrated by the test).

PiperOrigin-RevId: 347923407