Remove require_sequential_order option from BaseInputGenerator.
Move it into cluster instead. It really should be a property of some global
state rather than a property of any individual instance of a class.

With DataSource there can be many individual components, and it is difficult to
ensure that require_sequential_order is set consistently on all of them.

Note that after this change there will be no way to set require_sequential_order on a dataset from params. Thus, there will be just two modes:
- For unit tests, require_sequential_order will default to True, but individual tests can turn it off with a context manager
- For anything else, require_sequential_order will be None.

This should help reduce some amount of unit test flakiness due to input order.

PiperOrigin-RevId: 360789700