Add TFDatasetSource._map_args.
With deterministic=True, elements are output in deterministic order, but they
are not produced sequentially. If the map function internally contains a random
op, even if the seed is fixed, because of the parallel processing the results
can be nondeterministic.

Set num_parallel_calls=1 when in a unit test to work around this issue.

PiperOrigin-RevId: 361290977