When beam_search.batch_major_state=True, use corrected hyp ids to reorder along batch dim if and only if beam_search.batch_major_compute is True.
This is consistent with the behavior when batch_major_state=False.

Prior to this cl, old hyp ids are used regardless of batch_major_compute when batch_major_state=True. In other words, this cl changes the bahavior of ReOrderHyps(state) in beam search when batch_major_state=True and batch_major_state=True.

PiperOrigin-RevId: 369782140