Car input generator: Raise exception with extractor context.
Use Python3 syntax for re-raising an exception with specific
context, in this case, including the name of the extractor
that failed.

Otherwise you may get an unhelpful error message and stack trace that
doesn't pinpoint which extractor failed.  Once the user knows which
extractor failed, it is more likely for the error message to make
sense.

PiperOrigin-RevId: 370207596