Add force_eos_in_top_k param to beam_search_helper.
This param ensures the eos token is always included in top_k for each step, so
that termination of each hyp can be considered on every step.

Otherwise, if eos is just eps smaller than the k-th candidate, but still very
much within valid_eos_max_logit_delta and local_eos_threshold, it would not be
considered. If this happens for every time step, beam search returns no results,
because there are no terminated hyps found.

PiperOrigin-RevId: 377839313