Add Visit to hyperparams.
Visit implements a standard visitor pattern for hyperparams, so you
can build your own ToText or ToProto functions. This pattern is also
useful for modifying a sub-Params that's nested somewhere deep within
a big Params object such that it's hard to find.

I re-implemented ToText in terms of this pattern but didn't touch
ToProto because that looked more complicated and easier to break.

PiperOrigin-RevId: 380895901