Require that AQT weight names match their unquantized counterparts.
Currently it is impossible to tell which materialized AQT weight corresponds to a given `float32` layer weight. This makes the code more difficult to understand, makes it difficult to safeguard against a class of bugs, and makes it difficult to analyze the memory reduction that quantization provides for a given model.

This change ensures that the `w_name` that a user passes to `CreateAqtWeight`, `ToAqt*` and `FromAqt*` matches the `name` passed to `BaseLayer.CreateVariable`. The materialized AQT variables are stored under `f"layer.qdomain_default.{w_name}_aqt_weight"` and `f"layer.qdomain_default.{w_name}_aqt_scale"`.

Since old checkpoints may use the unstandardized variable names we need to provide an option to use the variable names currently at head until those checkpoints are updated or phased out. This is provided via `p.use_legacy_aqt_names`.

PiperOrigin-RevId: 391389229