Move calls to AddVN() into an AddGlobalVN() helper function.
When global_vn is enabled, AddGlobalVN() will be called in theta,
and will use the same noise vector for all timesteps (but a different
one for each global step).

When per_step_vn is enabled, self.AddVN(*, per_step=True) will be called inside
of the FProp of individual layers which results in a different noise vector for
each timestep. This requires deterministic=True.

Restore the default to deterministic=False.

PiperOrigin-RevId: 394364641