Remove writing of training graph GraphDef to log dirs.
Why:
1) For large, complicated models, converting the graph to graph_def
   to write out takes minutes.
2) For very large complicated models, the graph def exceeds the
   2GiB proto limit, and so the conversion fails.
3) It is not evident that these pbtxts are actually useful anymore.

At some point these training graphs were useful for debugging
and inspection, but tooling around these graph_defs is lacking.
It is unlikely that the training graphs themselves would ever be
used in isolation for debugging / reproducibility.

On the other hand, it is possible that the eval graphs could be useful,
so I have not yet removed them in this change.  It is also less likely
that the eval graph will have the same problems above as the training
graph, but we can make this change later if needed.

We have opted *not* to add an optional flag for this, because it is also
theoretically possible for the end-user to write out the graph themselves
if they need it.  The chance that a user really needed the train.pbtxt
to debug something is vanishingly small, compared to other source-code
archaeology and ML tracking tools that are available these days.

PiperOrigin-RevId: 396377280