Adding two options for cpu beam search.
1. force_last_chunk_eoc_in_top_k: if True, last chunk eoc will be forced into topk for an input hyp, so that the input hyp can be terminated by last chunk eoc when both eos and eoc scores are low.

With the current CL, the termination criteria for eos and eoc will mirror each other, except for using different tokens being added.

2. merged_topk_buffer_size_factor: by increasing this value to k+2, we disable early pruning and perform all possible path merges if merge_paths=True. This is useful for debugging the discrepancy between cpu beam search op vs. tpu beam search op (tpu beam search can not use a loop to keep shrinking the buffer size).

It is also possible that we were not able to get the most benefit from path merging (if the memory budget allows), due to its previous value 2 being a little bit low.

PiperOrigin-RevId: 402971620