Introduce _child_variable_scope_override.
Instead of acting on child objects, act on child names.

This is to prepare for moving to immediately creating variables in __init__. In
that case, the variable scopes need to be applied to CreateChild(ren) calls,
before the child objects are created.

This change needs to be performed in 3 steps:
1) (this cl) introduce _child_variable_scope_override
2) migrate most overrides of _CreateChildrenVariables to override
_child_variable_scope_override and/or _SelfVariableScope instead
3) apply _child_variable_scope_override and _SelfVariableScope to
CreateChild(ren) calls and remove _CreateChildrenVariables completely.

PiperOrigin-RevId: 416691416