Vectorize optimizers based on repeat_prefix
Uses a wrapper over existing gradient transformations to vectorize optimizers: the original optimizer code sees only variables without the repeat_prefix, then the wrapper uses vmap to add those dims in a parallel way. Since state and vars are not necessarily the same structure, we create per-group states, where each group consists of variables with the same prefixes.

This does not change checkpoint format if no repeat_prefix is used in any var.

If there is any var that uses repeat_prefix, the optimizer state will be a NestedMap with fields like:
{
  'no_prefix': state for vars without prefixes
  'p_2.3_0.-1': state for vars with shape prefix [2,3] and sharding prefix [0,-1]
  'p_4_1': state for vars with shape prefix [4] and sharding prefix [1]
   ...
}

PiperOrigin-RevId: 417732405