Previously, each executor program has its own Checkpointer for restoring. The logic is thus: 1) check if there are uninitialized variables. if not, assume everything is correct and return 2) if there are uninitialized variables, run the global initialization op, then check init_from_checkpoint_rules and override the variables from the rules
For multi-task models with two different init_from_checkpoint_rules however,
this logic breaks, because the second task sees everything as already
initialized and so its init_from_checkpoint_rules is not processed.

However, we cannot always blindly apply init_from_checkpoint_rules at the very
start of the program, because if training is resuming from a checkpoint we don't
want to apply init_from_checkpoint_rules again.

The best way to consolidate this is to have a single checkpointer for the model
rather than a separate one for each program. This is also more sensible because
there is a single checkpoint for the model, so there should only be a single
checkpointer to manage it.

PiperOrigin-RevId: 423412767