Add a de-duplicating AverageMetric.
Some input pipelines may not carefully ensure exactly-once
visitation of an evaluation set, leading to AverageMetric
sometimes over or under-counting the metric.  Because many
datasets have unique keys for each example, we can deduplicate
via the key (and also raise an error when inference
is non-deterministic for the same key) by adding a metric
implementation that tracks the key->value mapping and ignores
duplicates.

This should be used for cases where the size of this storage
is not prohibitively large, and where it is important to
know that there is no over or under counting.  This also
adds summaries reporting the total number of unique examples
that go into the average value, as well as the total weight
and the total value, if that is useful.

Adds a test of this behavior.

PiperOrigin-RevId: 424909961