Support unevenly partitioned vars.
The inputs/outputs of the init/train/eval/decode functions are padded to evenly shardable shapes, but inputs are immediately sliced at entry, and outputs will be padded before return, so that they will not affect the model code.

Checkpoints are still padded. TensorStore already allows reading a subset of previous data, so if sharding is changed to reduce padding it's still compatible. We will need to also allow out-of-bound TensorStore reads to support sharding changes that adds more padding.

PiperOrigin-RevId: 429612532