Added batch_mixing_partition_boundaries to BaseInputGeneratorFromFiles.
This parameter allows us to partition file_patterns when within_batch_mixing is off. All file_patterns in a partition are assigned to a single DataSource (instead of having a separate one for each file_pattern). This can significantly reduce the number of DataSources used, resulting in lower memory usage.

PiperOrigin-RevId: 432225054