Fixes setting of atten_logit_cap when trans_atten_tpl.atten_tpl is a list of attention classes.
Details:
batch_major_attention.TransformerAttentionLayer allows atten_tpl to be a list of attention classes. However the setting of atten_logit_cap in conformer_layer.ConformerLayer assumes that atten_tpl is not a list but a single attention class. This CL removes this assumption and handles a list of attention classes as well.
PiperOrigin-RevId: 434779651